+++ src/Command/BigStorage/Backup/Revert.php
@@ -2,11 +2,13 @@
 
 namespace Transip\Api\CLI\Command\BigStorage\Backup;
 
+use Transip\Api\CLI\Command\Field;
+use Transip\Api\CLI\Command\AbstractCommand;
+use Symfony\Component\Console\Input\ArrayInput;
+use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
-use Transip\Api\CLI\Command\AbstractCommand;
-use Transip\Api\CLI\Command\Field;
 
 class Revert extends AbstractCommand
 {
@@ -17,6 +19,8 @@ class Revert extends AbstractCommand
             ->addArgument(Field::BIGSTORAGE_NAME, InputArgument::REQUIRED, Field::BIGSTORAGE_NAME__DESC)
             ->addArgument(Field::BIGSTORAGE_BACKUPID, InputArgument::REQUIRED, Field::BIGSTORAGE_BACKUPID__DESC)
             ->addArgument(Field::BIGSTORAGE_BACKUP_DESTINATION_NAME, InputArgument::OPTIONAL, Field::BIGSTORAGE_BACKUP_DESTINATION_NAME__DESC . Field::OPTIONAL)
+            ->addOption(Field::ACTION_WAIT, 'w', InputOption::VALUE_NONE, Field::ACTION_WAIT_DESC)
+            ->addOption(Field::ACTION_PROGRESS, 'p', InputOption::VALUE_NONE, Field::ACTION_PROGRESS_DESC)
             ->setHelp('This command restores a big storage backup.');
     }
 
@@ -25,12 +29,35 @@ class Revert extends AbstractCommand
         $bigStorageName     = $input->getArgument(Field::BIGSTORAGE_NAME);
         $bigStorageBackupId = $input->getArgument(Field::BIGSTORAGE_BACKUPID);
         $destinationBigStorageName = $input->getArgument(Field::BIGSTORAGE_BACKUP_DESTINATION_NAME) ?? '';
+        $waitForAction = $input->getOption(Field::ACTION_WAIT);
+        $showProgress = $input->getOption(Field::ACTION_PROGRESS);
 
-        $this->getTransipApi()->bigStorageBackups()->revertBackup(
+        $response = $this->getTransipApi()->bigStorageBackups()->revertBackup(
             $bigStorageName,
             $bigStorageBackupId,
             $destinationBigStorageName
         );
+        $action = $this->getTransipApi()->actions()->parseActionFromResponse($response);
+
+
+        if ($action && $waitForAction) {
+            $app = $this->getApplication();
+            if (!$app) {
+                return 0;
+            }
+
+            $command = $app->get('action:pollstatus');
+            $arguments = [
+                'actionUuid'        => $action->getUuid()
+            ];
+
+            if ($showProgress) {
+                $arguments['--showProgress'] = true;
+            }
+    
+            $actionInput = new ArrayInput($arguments);
+            $command->run($actionInput, $output);
+        }
         return 0;
     }
 }