+++ src/Command/Vps/CloneVps.php
@@ -2,11 +2,13 @@
 
 namespace Transip\Api\CLI\Command\Vps;
 
+use Transip\Api\CLI\Command\Field;
+use Transip\Api\CLI\Command\AbstractCommand;
+use Symfony\Component\Console\Input\ArrayInput;
+use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
-use Transip\Api\CLI\Command\AbstractCommand;
-use Transip\Api\CLI\Command\Field;
 
 class CloneVps extends AbstractCommand
 {
@@ -16,6 +18,7 @@ class CloneVps extends AbstractCommand
             ->setDescription('Clone an existing VPS')
             ->addArgument(Field::VPS_NAME, InputArgument::REQUIRED, Field::VPS_NAME__DESC)
             ->addArgument(Field::AVAILABILITY_ZONE, InputArgument::OPTIONAL, Field::AVAILABILITY_ZONE__DESC . Field::OPTIONAL, '')
+            ->addOption(Field::ACTION_WAIT, 'w', InputOption::VALUE_NONE, Field::ACTION_WAIT_DESC)
             ->setHelp('You must provide the vps name of the VPS to clone, and optionally provide the name of the availability zone where the clone should be created');
     }
 
@@ -23,8 +26,27 @@ class CloneVps extends AbstractCommand
     {
         $vpsName = $input->getArgument(Field::VPS_NAME);
         $availabilityZone = $input->getArgument(Field::AVAILABILITY_ZONE);
+        $waitForAction = $input->getOption(Field::ACTION_WAIT);
 
-        $this->getTransipApi()->vps()->cloneVps($vpsName, $availabilityZone);
+        $response = $this->getTransipApi()->vps()->cloneVps($vpsName, $availabilityZone);
+        $action = $this->getTransipApi()->actions()->parseActionFromResponse($response);
+
+        if ($action && $waitForAction) {
+            $app = $this->getApplication();
+
+            if (!$app) {
+                return 0;
+            }
+
+            $command = $app->get('action:pollstatus');
+                
+            $arguments = [
+                'actionUuid'        => $action->getUuid()
+            ];
+
+            $actionInput = new ArrayInput($arguments);
+            $command->run($actionInput, $output);
+        }
         return 0;
     }
 }